delivery-bot/requirements.txt
fastapi
uvicorn
playwright
playwright-stealth
python-dotenv
pydantic
sqlalchemy
python-jose[cryptography]
requests
python-multipart
